requirements.txt
aerich==0.7.2
annotated-types==0.5.0
asyncpg==0.28.0
click==8.1.7
dictdiffer==0.9.0
greenlet==2.0.2
grpcio==1.58.0
grpcio-tools==1.58.0
iso8601==1.1.0
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
mypy-protobuf==3.5.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
protobuf==4.24.3
psycopg2-binary==2.9.7
pydantic==2.4.1
pydantic_core==2.10.1
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3.post1
tomlkit==0.12.1
tortoise-orm==0.20.0
types-protobuf==4.24.0.2
typing_extensions==4.8.0
tortoise-orm==0.20.0